+++ src/platform_impl/macos/window.rs
@@ -19,6 +19,7 @@ use crate::{
         LogicalPosition, LogicalSize, PhysicalPosition, PhysicalSize, Position, Size, Size::Logical,
     },
     error::{ExternalError, NotSupportedError, OsError as RootOsError},
+    event::WindowEvent,
     icon::Icon,
     platform::macos::WindowExtMacOS,
     platform_impl::platform::{
@@ -32,8 +33,8 @@ use crate::{
         Fullscreen, OsError,
     },
     window::{
-        CursorGrabMode, CursorIcon, ResizeDirection, Theme, UserAttentionType, WindowAttributes,
-        WindowButtons, WindowId as RootWindowId, WindowLevel,
+        CursorGrabMode, CursorIcon, ImePurpose, ResizeDirection, Theme, UserAttentionType,
+        WindowAttributes, WindowButtons, WindowId as RootWindowId, WindowLevel,
     },
 };
 use core_graphics::display::{CGDisplay, CGPoint};
@@ -216,7 +217,7 @@ impl WinitWindow {
         }
 
         let this = autoreleasepool(|_| {
-            let screen = match &attrs.fullscreen {
+            let screen = match attrs.fullscreen.clone().map(Into::into) {
                 Some(Fullscreen::Borderless(Some(monitor)))
                 | Some(Fullscreen::Exclusive(VideoMode { monitor, .. })) => {
                     monitor.ns_screen().or_else(NSScreen::main)
@@ -458,19 +459,25 @@ impl WinitWindow {
         let delegate = WinitWindowDelegate::new(&this, attrs.fullscreen.is_some());
 
         // Set fullscreen mode after we setup everything
-        this.set_fullscreen(attrs.fullscreen);
+        this.set_fullscreen(attrs.fullscreen.map(Into::into));
 
         // Setting the window as key has to happen *after* we set the fullscreen
         // state, since otherwise we'll briefly see the window at normal size
         // before it transitions.
         if attrs.visible {
-            // Tightly linked with `app_state::window_activation_hack`
-            this.makeKeyAndOrderFront(None);
+            if attrs.active {
+                // Tightly linked with `app_state::window_activation_hack`
+                this.makeKeyAndOrderFront(None);
+            } else {
+                this.orderFront(None);
+            }
         }
 
         if attrs.maximized {
             this.set_maximized(attrs.maximized);
         }
+
+        delegate.queue_event(WindowEvent::Focused(false));
 
         Ok((this, delegate))
     }
@@ -1146,6 +1153,9 @@ impl WinitWindow {
         // TODO(madsmtm): Remove the need for this
         unsafe { Id::from_shared(self.view()) }.set_ime_allowed(allowed);
     }
+
+    #[inline]
+    pub fn set_ime_purpose(&self, _purpose: ImePurpose) {}
 
     #[inline]
     pub fn focus_window(&self) {